requirements.txt
pdfminer.six==20201018
lxml
numpy
pillow